requirements.txt
pytest==8.3.3
behave==1.2.6
playwright==1.48.0
allure-pytest==2.13.5
allure-behave==2.13.5
websocket-client==1.8.0
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
